* jl/nppcr-precompiled-headers-and-w4-wx:
  Forgot one #include <windows.h>
  Introduced precompiled headers.
  Started the precompiled header stuff.
  Final Warning cleanup.
  Warning Cleanup, phase 5.
  Warning Cleanup, Phase 4.
  Warning Cleanup, Phase 3.
  "Really" fix all the warnings for Phase 1.
  Warning Cleanup, Phase 2.
  Warning cleanup, Phase 1.
  Removed most of the warnings for Debug Unicode.
  First pass cleanup.

+++ PowerEditor/src/ScitillaComponent/FunctionCallTip.cpp
@@ -15,6 +15,7 @@
 //along with this program; if not, write to the Free Software
 //Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 
+#include "precompiled_headers.h"
 #include "FunctionCallTip.h"
 
 
@@ -157,14 +158,14 @@ bool FunctionCallTip::getCursorFunction() {
 		}
 	}
 
-	size_t vsize = tokenVector.size();
+	int vsize = (int)tokenVector.size();
 	//mind nested funcs, like |blblb a (x, b(), c);|
 	//therefore, use stack
 	std::vector<FunctionValues> valueVec;
 
 	FunctionValues curValue, newValue;
 	int scopeLevel = 0;
-	for (size_t i = 0; i < vsize; i++) {
+	for (int i = 0; i < vsize; i++) {
 		Token & curToken = tokenVector.at(i);
 		if (curToken.isIdentifier) {
 			curValue.lastIdentifier = i;
@@ -342,8 +343,8 @@ void FunctionCallTip::showCalltip() {
 	if (hasDescr)
 		bytesNeeded += lstrlen(curDescriptionText);
 
-	size_t nrParams = params.size();
-	for(size_t i = 0; i < nrParams; i++) {
+	int nrParams = (int)params.size();
+	for(int i = 0; i < nrParams; i++) {
 		bytesNeeded += lstrlen(params.at(i)) + 2;	//'param, '
 	}
 
@@ -369,7 +370,7 @@ void FunctionCallTip::showCalltip() {
 
 	int highlightstart = 0;
 	int highlightend = 0;
-	for(size_t i = 0; i < nrParams; i++) {
+	for(int i = 0; i < nrParams; i++) {
 		if (i == _currentParam) {
 			highlightstart = lstrlen(textBuffer);
 			highlightend = highlightstart + lstrlen(params.at(i));
